requirements.txt
asgiref==3.11.0
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
coverage==7.13.4
Django==5.2.9
django-debug-toolbar==6.2.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==39.0.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
parameterized==0.9.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
pytest==9.0.2
pytest-django==4.11.1
pytest-dotenv==0.5.2
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.4
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-PyYAML==6.0.12.20250915
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
